modified:   ants_on_table.cc
   float -> double for constant "#defines" to maintain original rounding.

+++ phy1610/ps2/ants_on_table.cc
@@ -64,9 +64,9 @@ float ants_on_table::total_number_of_ants() const
 
 void ants_on_table::simulation( const int number_of_time_intervals )
 {
-   static const float RADIUS_OF_POSITION_ADJUSTMENT { 1.9 } ;
-   static const float NOT_WALKING_ANTS_FRACTION { 0.8 } ;
-   static const float FALLEN_ANTS_FRACTION { 0.2 } ;
+   static const double RADIUS_OF_POSITION_ADJUSTMENT { 1.9 } ;
+   static const double NOT_WALKING_ANTS_FRACTION { 0.8 } ;
+   static const double FALLEN_ANTS_FRACTION { 0.2 } ;
 
    rarray<float, 2> new_number_of_ants( m_table_grid_size, m_table_grid_size ) ;
